littlelemonAPI/LittleLemon/requirements.txt
Django>=4.2
djangorestframework>=3.14
djoser>=2.2